Rename class directory
