+++ app/scripts.babel/classes/HasValue.js
