+++ app/scripts.babel/classes/Message.js
