+++ app/scripts.babel/classes/Talk.js
