+++ app/scripts.babel/classes/TalkArea.js
